Refactored Post endpoints for authentication

+++ app/Http/Controllers/Post/PostController.php
@@ -11,23 +11,16 @@ use App\Http\Requests\PostUpdateRequest;
 
 class PostController extends ApiController
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    
     public function index()
     {
-        $posts=Post::all();
+        $posts=Post::all()
+                   ->load('image')
+                   ->load('tags');
         return $this->showCollectionAsResponse($posts);
-
     }
 
-    /**
-     * Returns the tag id's of tags passed in
-     * Array of tags passed in
-     * Returns the ids of those tags
-     */
+    
     private function creatOrFetchTagIds($taglist)
     {
         $tags=[];
@@ -43,16 +36,11 @@ class PostController extends ApiController
         return $tags;
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+  
     public function store(PostStoreRequest $request)
     {
         $data=$request->only(['title','body']);
-        $data['user_id']=1; //to be replace with id of current logged in user
+        $data['user_id']=auth()->user()->id;
         $post=Post::create($data);
         
         if($request->has('tags'))
@@ -70,32 +58,27 @@ class PostController extends ApiController
         return $this->showModelAsResponse($post,201);    
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Post  $post
-     * @return \Illuminate\Http\Response
-     */
+    
     public function show(Post $post)
     {
         return $this->showModelAsResponse($post);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Post  $post
-     * @return \Illuminate\Http\Response
-     */
+  
     public function update(PostUpdateRequest $request, Post $post)
     {
+        
+        if($post->user->id!=auth()->user()->id)
+        {
+            return $this->errorResponse("You don't have permission to update this post",409);   
+        }
+
         $post->update($request->only(['title','body']));
 
         if($request->has('tags'))
         {
             $tags=$this->creatOrFetchTagIds($request->tags);
-            $post->tags()->sync($tags);
+            $post->tags()->syncWithoutDetaching($tags);
         }
 
         if($request->hasFile('image'))
@@ -107,16 +90,15 @@ class PostController extends ApiController
         return $this->showModelAsResponse($post);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Post  $post
-     * @return \Illuminate\Http\Response
-     */
+   
     public function destroy(Post $post)
     {
+        if($post->user->id!=auth()->user()->id)
+        {
+            return $this->errorResponse("You don't have permission to delete this post",409);   
+        }
+
         $post->delete();
-        
         return $this->showModelAsResponse($post);
     }
 }
